Desktop Overhaul
Badge and BadgeAwarded now store IDs instead of pointing to eachother
BadgeAwarded is granted for each Badge in the database and defaulted to 0 (Inprogress)

Removed unused functions

Added check to ensure the user is in the database before alowing login
-You can add your fedid email on the login screen to allow access

A complete overhaul of views. Different views can be invoked by clearing the main container and inserting a new widget.
See Whoseturn.java for these changes and see all the new views in edu.unlv.cs.whoseturn.views.desktop